Object detection (#2)
* Requirements file now points to latest gym-utils and _vgdl

Signed-off-by: Jon Crawford <jon@joncrawf.com>

* Use contours for object detection with clustering for objects classes

When objdetect option is selected, the environment is wrapped by a
object detection wrapper. This wrapper uses contours to detect the
objects, which then are clustered to define their object class. Another
object attribute is given through its velocity which is calculated
through the nearest object with the same class from the previous frame.

Signed-off-by: Jon Crawford <jon@joncrawf.com>